Specify grant_type only in body of request to Google
Recently, users could not login using Google OAuth2.  The error in the
logs was:

    org.scribe.exceptions.OAuthException: Cannot extract an acces token.
    Response was: {
      "error" : "invalid_request",
        "error_description" : "OAuth 2 parameters can only have a single
        value: grant_type"
        }

Perhaps it is a recent change that Google does not accept duplicate
parameters to appear in both URL and body of the request.  This
thread seems to indicate similar behavior for other parameters.

    http://stackoverflow.com/questions/29694840/googles-openid-connect-says-oauth-2-parameters-can-only-have-a-single-value-cuo

Remove the grant_type parameter from the URL, since it is specified
in the body of the request.